Explicitly pass -XCPP to hlint
This is necessary because haskell-ci now passes -XHaskell2010 to hlint, which
causes it to disable language extensions that aren't part of Haskell2010 (e.g.,
CPP). This has the unfortunate consequence of hlint failing to parse any
modules that use CPP. To work around this, we now pass -XCPP explicitly to
hlint. This is a bit unfortunate—see ndmitchell/hlint#1360—but this should
let the CI pass for now.